+++ service/controller/resource/machinepoolupgrade/lastdeployedreleaseversion.go
@@ -9,6 +9,7 @@ import (
 	capiexp "sigs.k8s.io/cluster-api/exp/api/v1alpha3"
 	"sigs.k8s.io/cluster-api/util"
 
+	"github.com/giantswarm/azure-operator/v5/pkg/tenantcluster"
 	"github.com/giantswarm/azure-operator/v5/pkg/upgrade"
 	"github.com/giantswarm/azure-operator/v5/service/controller/key"
 )
@@ -35,7 +36,12 @@ func (r *Resource) ensureLastDeployedReleaseVersion(ctx context.Context, machine
 	}
 
 	tenantClusterClient, err := r.tenantClientFactory.GetClient(ctx, cluster)
-	if err != nil {
+	if tenantcluster.IsAPINotAvailableError(err) {
+		r.logger.LogCtx(ctx, "level", "debug", "message", "tenant API not available yet")
+		r.logger.LogCtx(ctx, "level", "debug", "message", "canceling resource")
+
+		return nil
+	} else if err != nil {
 		return microerror.Mask(err)
 	}
 	upgradeIsCompletedForDesiredVersion, err := upgrade.IsNodePoolUpgradeCompleted(